localstack/requirements.txt
awscli-local[ver1]
deepdiff
jinja2
junit-xml
localstack
PyYAML
terraform-local
